PUT request "/shopItem" (#11)
This closes #10 

**PUT request to update item name:**

`shopItemById.js`

Looks like my code doesn't work. I have tried to test it with [https://reqbin.com/curl], however the response was "error sending request".
I also tried sending requests in command line but did not work either.
I left some logs in the put request, sorry for that!
 
First, I check if the type of the query is not a number, and then I send back a 400 status. 
- I was thinking if this logic makes the issue, as the req.query comes like an Object {"id": "number"}, but I haven't tested it out.

Second, I look for the index of the `req.query.ID`. If we don't have it in our database, I send back a 404 status.

If we have it, I overwrite the value of that specific objects name with the name the user sent. 

the I rewrite the database and send back a 200 status.